Charles Ives Tribute Page
This is my first attempt at really designing a simple webpage from scratch using HTML, CSS, and Bootstrap. It still needs some work and I'm finishing it up as there are some tweaks that need to be made to the formatting. I'm sure that that code needs to be cleaned up a little too because I have this feeling that it is kind of a mess right now. A work in progress, but nearly done.